Scalable MPI implementation

+++ main.cpp
@@ -1,4 +1,4 @@
-#include <string>
+#include <string.h>
 #include <iostream>
 #include <vector>
 #include <memory>
@@ -10,7 +10,6 @@
 #include "fileManager.h"
 
 #include "mpi.h"
-#include "omp.h"
 
 using namespace std;
 
@@ -27,26 +26,77 @@ int main(int argc, char *argv[])
 	MPI_Comm_size(MPI_COMM_WORLD, &numprocs);
 	MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 	MPI_Get_processor_name(processor_name, &namelen);
-
-	fileManager fileHandler;
-	joinManager test;
-		
-	fileHandler.deleteFile("joinedFile.txt");
 	
-	struct timeval  tv1, tv2;
+	struct timeval tv1, tv2;
     gettimeofday(&tv1, NULL);
 
-	vector<string> linesOfFile1 = fileHandler.readFile("smallInput1.txt",1);
-	vector<string> linesOfFile2 = fileHandler.readFile("smallInput2.txt",1);
+	printf("Working on process %d out of %d on %s\n", rank, numprocs, processor_name);
 
-	#pragma omp parallel default(shared) private(iam, np)
-	{
-		np = omp_get_num_threads();
-    	iam = omp_get_thread_num();
-		printf("Working on thread %d out of %d from process %d out of %d on %s\n",
-           iam, np, rank, numprocs, processor_name);
+	
+	if (rank == 0) {
+		fileManager fileHandler;
+		fileHandler.deleteFile("joinedFile.txt");
 
-		#pragma omp for
+		vector<string> linesOfFile1, linesOfFile2;
+
+		linesOfFile1 = fileHandler.readFile("smallInput1.txt",1);
+		linesOfFile2 = fileHandler.readFile("smallInput2.txt",1);
+
+		int length_file_1 = linesOfFile1.size();
+		int length_file_2 = linesOfFile2.size();
+
+		for (auto i = 1; i < numprocs; i++){
+			int process_length_file_2 = length_file_2 / (numprocs - 1.0);
+			MPI_Send(&length_file_1, 1, MPI_INT, i, 0, MPI_COMM_WORLD);
+			MPI_Send(&process_length_file_2, 1, MPI_INT, i, 0, MPI_COMM_WORLD);
+
+			for (auto x=0; x < length_file_1; x++){
+				const char* line_ptr = linesOfFile1[x].c_str();
+				string line = linesOfFile1[x].c_str();
+				int line_length = line.length();
+				MPI_Send(&line_length, 1, MPI_INT, i, 0, MPI_COMM_WORLD);
+				MPI_Send(line_ptr, line_length, MPI_CHAR, i, 0, MPI_COMM_WORLD);
+			}
+
+			for (auto y= (i-1) * process_length_file_2; y < i * process_length_file_2; y++){
+				const char* line_ptr = linesOfFile2[y].c_str();
+				string line = linesOfFile2[y].c_str();
+				int line_length = line.length();
+				MPI_Send(&line_length, 1, MPI_INT, i, 0, MPI_COMM_WORLD);
+				MPI_Send(line_ptr, line_length, MPI_CHAR, i, 0, MPI_COMM_WORLD);
+			}
+		}
+	}
+	else {
+		int length_file_1, length_file_2;
+		vector<string> linesOfFile1, linesOfFile2;
+		joinManager test;
+
+		MPI_Recv(&length_file_1, 1, MPI_INT, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+		MPI_Recv(&length_file_2, 1, MPI_INT, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+
+		cout << "File 2 length: " << length_file_2 << endl;
+
+		for (auto i=0; i < length_file_1; i++){
+			int line_length;
+			char char_array[1024];
+			memset(char_array, 0, sizeof(char_array));
+			MPI_Recv(&line_length, 1, MPI_INT, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+			MPI_Recv(&char_array[0], line_length, MPI_CHAR, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+			string line(char_array);
+			linesOfFile1.push_back(line);
+		}
+
+		for (auto i=0; i < length_file_2; i++){
+			int line_length;
+			char char_array[1024];
+			memset(char_array, 0, sizeof(char_array));
+			MPI_Recv(&line_length, 1, MPI_INT, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+			MPI_Recv(&char_array[0], line_length, MPI_CHAR, 0, 0, MPI_COMM_WORLD,MPI_STATUS_IGNORE);
+			string line(char_array);
+			linesOfFile2.push_back(line);
+		}
+
 		for (unsigned int i=0; i< linesOfFile1.size(); i = i+2)
 		{
 			//int tid = omp_get_thread_num();  
@@ -55,7 +105,7 @@ int main(int argc, char *argv[])
 			string value = linesOfFile1[i+1];
 			test.hasher -> AddItem(key,value);
 		}
-		#pragma omp for
+
 		for (unsigned int i=0; i<linesOfFile2.size();i = i+2)
 		{
 		//	int tid = omp_get_thread_num();  
@@ -72,5 +122,4 @@ int main(int argc, char *argv[])
     printf("\n");
 
 	MPI_Finalize();
-	
 }
